Fix todo list and daily log problems
Issue 1: To-do list not staying "checked"
The problem is that our loading completion states in multiple places inconsistently. When we log emotions, the real-time listener updates weekLogs but doesn't call loadCompletionStates() which updates the to-do list states.

Issue 2: Deleting one emotion deletes all, then they reappear
This is because EmotionLog.id is null (not populated from Firestore). When we delete, we're passing log.id which is null, causing Firestore to potentially delete incorrectly or not find the document.